+++ crates/deed-ast/src/lib.rs
@@ -441,7 +441,7 @@ impl Stmt {
     }
 }
 
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
 pub enum BinaryOp {
     Or,
     And,